// client/stocklinx app/src/redux/permission/constant.ts
export enum permissionConst {
  FETCH_PERMISSIONS_REQUEST = "FETCH_PERMISSIONS_REQUEST",
  FETCH_PERMISSIONS_SUCCESS = "FETCH_PERMISSIONS_SUCCESS",
  FETCH_PERMISSIONS_FAILURE = "FETCH_PERMISSIONS_FAILURE",
  FETCH_PERMISSION_REQUEST = "FETCH_PERMISSION_REQUEST",
  FETCH_PERMISSION_SUCCESS = "FETCH_PERMISSION_SUCCESS",
  FETCH_PERMISSION_FAILURE = "FETCH_PERMISSION_FAILURE",
  FETCH_PERMISSION_COUNTS_REQUEST = "FETCH_PERMISSION_COUNTS_REQUEST",
  FETCH_PERMISSION_COUNTS_SUCCESS = "FETCH_PERMISSION_COUNTS_SUCCESS",
  FETCH_PERMISSION_COUNTS_FAILURE = "FETCH_PERMISSION_COUNTS_FAILURE",
  CREATE_PERMISSION_REQUEST = "CREATE_PERMISSION_REQUEST",
  CREATE_PERMISSION_SUCCESS = "CREATE_PERMISSION_SUCCESS",
  CREATE_PERMISSION_FAILURE = "CREATE_PERMISSION_FAILURE",
  CREATE_RANGE_PERMISSION_REQUEST = "CREATE_RANGE_PERMISSION_REQUEST",
  CREATE_RANGE_PERMISSION_SUCCESS = "CREATE_RANGE_PERMISSION_SUCCESS",
  CREATE_RANGE_PERMISSION_FAILURE = "CREATE_RANGE_PERMISSION_FAILURE",
  REMOVE_PERMISSION_REQUEST = "REMOVE_PERMISSION_REQUEST",
  REMOVE_PERMISSION_SUCCESS = "REMOVE_PERMISSION_SUCCESS",
  REMOVE_PERMISSION_FAILURE = "REMOVE_PERMISSION_FAILURE",
  REMOVE_RANGE_PERMISSION_REQUEST = "REMOVE_RANGE_PERMISSION_REQUEST",
  REMOVE_RANGE_PERMISSION_SUCCESS = "REMOVE_RANGE_PERMISSION_SUCCESS",
  REMOVE_RANGE_PERMISSION_FAILURE = "REMOVE_RANGE_PERMISSION_FAILURE",
  SYNC_PERMISSION_REQUEST = "SYNC_PERMISSION_REQUEST",
  SYNC_PERMISSION_SUCCESS = "SYNC_PERMISSION_SUCCESS",
  SYNC_PERMISSION_FAILURE = "SYNC_PERMISSION_FAILURE",
  FILTER_PERMISSIONS_REQUEST = "FILTER_PERMISSIONS_REQUEST",
  FILTER_PERMISSIONS_SUCCESS = "FILTER_PERMISSIONS_SUCCESS",
  FILTER_PERMISSIONS_FAILURE = "FILTER_PERMISSIONS_FAILURE",
  SET_PERMISSION = "SET_PERMISSION",
  CLEAR_PERMISSION = "CLEAR_PERMISSION",
  SET_PERMISSIONS = "SET_PERMISSIONS",
  CLEAR_PERMISSIONS = "CLEAR_PERMISSIONS",
}